pyproject.toml
[project]
name = "mrc-sop-app"
version = "0.1.0"
description = "Mulligan Recovery Centers SOP portal"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}

dependencies = [
  "fastapi==0.111.0",
  "uvicorn==0.30.1",
  "jinja2==3.1.4",
  "python-docx==1.1.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
